requirements.txt
dbt-core
dbt-snowflake
snowflake-connector-python[pandas]
pandas
pyarrow
python-dotenv
reverse_geocoder
fastapi
uvicorn
python-multipart
